cybergrid-controller/requirements.txt
Flask>=3.0.0,<4.0.0
Flask-Cors>=4.0.0,<5.0.0
SQLAlchemy>=2.0.0,<3.0.0
passlib>=1.7.4,<2.0.0
PyJWT>=2.8.0,<3.0.0
cryptography>=42.0.0,<43.0.0

